# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: PBS Service
  version: 1.0.0
servers:
  - url: 'https://localhost/pbs-service'
paths:
  /files-upload:
    post:
      tags:
        - Files
      summary: API for upload images
      operationId: uploadFiles
      requestBody:
        content:
          image/jpg:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Upload successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadFileResponse'
    get:
      tags:
        - Files
      summary: API for get images
      operationId: getAllFiles
      responses:
        '200':
          description: get successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFileResponse'
    put:
      tags:
        - Files
      summary: API for update images
      operationId: updateFile
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UploadFileResponse'
      responses:
        '200':
          description: Delete successfully
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/UploadFileResponse'
        '400':
          description: Bad request
  /files-upload/{id}:
    delete:
      tags:
        - Files
      summary: API for delete images
      operationId: deleteFile
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: image ID.
      responses:
        '200':
          description: Delete successfully
        '400':
          description: Bad request
  /files-upload/{id}/tags:
    put:
      tags:
        - Files
      summary: API for create tag or update tags
      operationId: putTags
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: image ID.
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/PutTagsRequest'
      responses:
        '200':
          description: Put successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadFileResponse'
        '400':
          description: Bad request

components:
  schemas:
    UploadFileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileUrl:
          type: string
        tags:
          type: string
        character:
          type: string
    ListFileResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UploadFileResponse'
    DeleteFileRequest:
      type: object
      properties:
        url:
          type: string
    PutTagsRequest:
      type: object
      properties:
        tags:
          type: string